make torch optional for plotting #54
